+++ Cargo.toml
@@ -464,6 +464,7 @@ revm-inspector = { version = "8.0.2", default-features = false }
 revm-context = { version = "8.0.2", default-features = false }
 revm-context-interface = { version = "8.0.1", default-features = false }
 revm-database-interface = { version = "7.0.1", default-features = false }
+revm-precompile = { version = "24.0.1", default-features = false }
 op-revm = { version = "8.0.3", default-features = false }
 revm-inspectors = "0.27.1"
 
@@ -714,6 +715,8 @@ vergen-git2 = "1.0.5"
 
 [patch.crates-io]
 alloy-eip2124 = { git = "https://github.com/alloy-rs/eips", branch = "rkrasiuk/fork-hash-from-bytes-arr" }
+alloy-evm = { git = "https://github.com/alloy-rs/evm", branch = "fusaka-devnet-3" }
+alloy-op-evm = { git = "https://github.com/alloy-rs/evm", branch = "fusaka-devnet-3" }
 
 alloy-consensus = { git = "https://github.com/alloy-rs/alloy", branch = "fusaka-devnet-3-tmp" }
 alloy-contract = { git = "https://github.com/alloy-rs/alloy", branch = "fusaka-devnet-3-tmp" }
@@ -753,6 +756,7 @@ revm-inspector = { git = "https://github.com/bluealloy/revm", branch = "fusaka-d
 revm-context = { git = "https://github.com/bluealloy/revm", branch = "fusaka-devnet-3" }
 revm-context-interface = { git = "https://github.com/bluealloy/revm", branch = "fusaka-devnet-3" }
 revm-database-interface = { git = "https://github.com/bluealloy/revm", branch = "fusaka-devnet-3" }
+revm-precompile = { git = "https://github.com/bluealloy/revm", branch = "fusaka-devnet-3" }
 op-revm = { git = "https://github.com/bluealloy/revm", branch = "fusaka-devnet-3" }
 
 # op-alloy-consensus = { git = "https://github.com/alloy-rs/op-alloy", rev = "a79d6fc" }
