+++ examples/custom-beacon-withdrawals/src/main.rs
@@ -155,6 +155,10 @@ impl ConfigureEvm for CustomEvmConfig {
     ) -> EthBlockExecutionCtx<'_> {
         self.inner.context_for_next_block(parent, attributes)
     }
+
+    fn precompiles_for_timestamp(&self, timestamp: u64) -> PrecompilesMap {
+        self.inner.precompiles_for_timestamp(timestamp)
+    }
 }
 
 pub struct CustomBlockExecutor<'a, Evm> {
