+++ examples/custom-node/src/evm/config.rs
@@ -80,4 +80,8 @@ impl ConfigureEvm for CustomEvmConfig {
             extra_data: attributes.extra_data,
         }
     }
+
+    fn precompiles_for_timestamp(&self, timestamp: u64) -> alloy_evm::precompiles::PrecompilesMap {
+        self.inner.precompiles_for_timestamp(timestamp)
+    }
 }
